requirements.txt
ultralytics
opencv-python
supervision
numpy
pyyaml
flask
flask-cors
